requirements.txt
libzim==3.4.0
trafilatura==1.10.0
lxml==5.2.2
sentence-transformers==3.0.0
faiss-cpu==1.8.0